fix: gTag script location _app to _document
