+++ src/pages/_app.tsx
@@ -3,11 +3,8 @@ import '@/styles/global.css'
 import Nav from '@/components/nav'
 import Head from 'next/head'
 import Footer from '@/components/footer'
-import Script from 'next/script'
 
 export default function App({ Component, pageProps }: AppProps) {
-  const GA_TRACKING_ID = process.env.NEXT_PUBLIC_GA_TRACKING_ID
-  const GTM_ID = process.env.NEXT_PUBLIC_GTM_ID
   return (
     <>
       <Head>
@@ -19,61 +16,7 @@ export default function App({ Component, pageProps }: AppProps) {
         <meta property="og:type" content="website" />
         <meta name="theme-color" content="#F0E5DE" />
       </Head>
-      <Script
-        src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`}
-        strategy="afterInteractive"
-      />
-      <Script
-        id="google-analytics"
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', '${GA_TRACKING_ID}', {
-              page_path: window.location.pathname,
-            });
-          `,
-        }}
-      />
-      <Script
-        id="gtm"
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-            (function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
-            new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
-            j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
-            'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-            })(window,document,'script','dataLayer','${GTM_ID}');
-          `,
-        }}
-      />
-      <Script
-        id="facebook-sdk"
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-            window.fbAsyncInit = function() {
-              FB.init({
-                appId      : '1254747732413908',
-                xfbml      : true,
-                version    : 'v21.0'
-              });
-              FB.AppEvents.logPageView();
-            };
 
-            (function(d, s, id){
-               var js, fjs = d.getElementsByTagName(s)[0];
-               if (d.getElementById(id)) {return;}
-               js = d.createElement(s); js.id = id;
-               js.src = "https://connect.facebook.net/en_US/sdk.js";
-               fjs.parentNode.insertBefore(js, fjs);
-             }(document, 'script', 'facebook-jssdk'));
-          `,
-        }}
-      />
       <Nav />
       <Component {...pageProps} />
       <Footer />
